@@ -8,10 +8,10 @@ pub mod message;
 pub mod processor;
 pub mod receiver;
 
+pub mod config;
+mod effect_handler;
+pub mod local;
+pub mod pipeline;
+pub mod shared;
 #[cfg(test)]
 pub mod testing;
-
-use std::rc::Rc;
-
-/// A type representing the name of a node in the pipeline.
-pub type NodeName = Rc<str>;
