@@ -1,41 +1,52 @@
 // SPDX-License-Identifier: Apache-2.0
 
-//! Utility functions for testing receivers.
-//!
-//! This module provides specialized utilities for testing receiver components:
-//!
-//! - `ReceiverTestContext`: Provides a context for interacting with receivers during tests
-//! - `ReceiverTestRuntime`: Configures and manages a single-threaded tokio runtime for receiver tests
+//! Test utilities for receivers.
 //!
 //! These utilities are designed to make testing receivers simpler by abstracting away common
 //! setup and lifecycle management.
 
-use crate::message::ControlMsg;
-use crate::receiver::{ControlMsgChannel, Receiver};
-use crate::testing::{CtrMsgCounters, TestMsg, create_test_channel, setup_test_runtime};
-use otap_df_channel::error::SendError;
-use otap_df_channel::mpsc;
+use crate::config::ReceiverConfig;
+use crate::error::Error;
+use crate::message::{ControlMsg, Receiver, Sender};
+use crate::receiver::ReceiverWrapper;
+use crate::testing::{CtrlMsgCounters, setup_test_runtime};
+use otap_df_channel::error::RecvError;
 use serde_json::Value;
+use std::fmt::Debug;
+use std::marker::PhantomData;
 use std::time::Duration;
 use tokio::task::LocalSet;
 use tokio::time::sleep;
 
-/// A context object for receiver tests that provides functionality for interacting with receivers.
-pub struct ReceiverTestContext {
+/// Context used during the test phase of a test.
+pub struct TestContext {
     /// Sender for control messages
-    control_tx: mpsc::Sender<ControlMsg>,
-    /// Receiver for messages sent by the receiver
-    pdata_rx: Option<mpsc::Receiver<TestMsg>>,
+    control_sender: Sender<ControlMsg>,
 }
 
-impl ReceiverTestContext {
+/// Context used during the validation phase of a test (!Send context).
+pub struct NotSendValidateContext<PData> {
+    pdata_receiver: Receiver<PData>,
+    counters: CtrlMsgCounters,
+}
+
+/// Context used during the validation phase of a test (Send context).
+pub struct SendValidateContext<PData> {
+    pdata_receiver: tokio::sync::mpsc::Receiver<PData>,
+    counters: CtrlMsgCounters,
+}
+
+impl TestContext {
     /// Sends a timer tick control message.
     ///
     /// # Errors
     ///
     /// Returns an error if the message could not be sent.
-    pub async fn send_timer_tick(&self) -> Result<(), SendError<ControlMsg>> {
-        self.control_tx.send_async(ControlMsg::TimerTick {}).await
+    pub async fn send_timer_tick(&self) -> Result<(), Error<ControlMsg>> {
+        self.control_sender
+            .send(ControlMsg::TimerTick {})
+            .await
+            .map_err(Error::ChannelSendError)
     }
 
     /// Sends a config control message.
@@ -43,10 +54,11 @@ impl ReceiverTestContext {
     /// # Errors
     ///
     /// Returns an error if the message could not be sent.
-    pub async fn send_config(&self, config: Value) -> Result<(), SendError<ControlMsg>> {
-        self.control_tx
-            .send_async(ControlMsg::Config { config })
+    pub async fn send_config(&self, config: Value) -> Result<(), Error<ControlMsg>> {
+        self.control_sender
+            .send(ControlMsg::Config { config })
             .await
+            .map_err(Error::ChannelSendError)
     }
 
     /// Sends a shutdown control message.
@@ -54,22 +66,50 @@ impl ReceiverTestContext {
     /// # Errors
     ///
     /// Returns an error if the message could not be sent.
-    pub async fn send_shutdown(&self, reason: &str) -> Result<(), SendError<ControlMsg>> {
-        self.control_tx
-            .send_async(ControlMsg::Shutdown {
+    pub async fn send_shutdown(
+        &self,
+        deadline: Duration,
+        reason: &str,
+    ) -> Result<(), Error<ControlMsg>> {
+        self.control_sender
+            .send(ControlMsg::Shutdown {
+                deadline,
                 reason: reason.to_owned(),
             })
             .await
+            .map_err(Error::ChannelSendError)
     }
 
     /// Sleeps for the specified duration.
     pub async fn sleep(&self, duration: Duration) {
         sleep(duration).await;
     }
+}
 
-    /// Returns a reference to the control message sender.
-    pub fn pdata_rx(&mut self) -> Option<mpsc::Receiver<TestMsg>> {
-        self.pdata_rx.take()
+impl<PData> NotSendValidateContext<PData> {
+    /// Receives a pdata message produced by the receiver.
+    pub async fn recv(&mut self) -> Result<PData, RecvError> {
+        self.pdata_receiver.recv().await
+    }
+
+    /// Returns the control message counters.
+    pub fn counters(&self) -> CtrlMsgCounters {
+        self.counters.clone()
+    }
+}
+
+impl<PData> SendValidateContext<PData> {
+    /// Receives a pdata message produced by the receiver.
+    pub async fn recv(&mut self) -> Result<PData, Error<PData>> {
+        self.pdata_receiver
+            .recv()
+            .await
+            .ok_or(Error::ChannelRecvError(RecvError::Closed))
+    }
+
+    /// Returns the control message counters.
+    pub fn counters(&self) -> CtrlMsgCounters {
+        self.counters.clone()
     }
 }
 
@@ -77,88 +117,122 @@ impl ReceiverTestContext {
 ///
 /// This structure encapsulates the common setup logic needed for testing receivers,
 /// including channel creation, receiver instantiation, and task management.
-pub struct ReceiverTestRuntime {
+pub struct TestRuntime<PData> {
+    /// The configuration for the receiver
+    config: ReceiverConfig,
+
     /// Runtime instance
     rt: tokio::runtime::Runtime,
     /// Local task set for non-Send futures
     local_tasks: LocalSet,
 
-    /// Sender for control messages
-    control_tx: mpsc::Sender<ControlMsg>,
-    /// Receiver for control messages
-    control_rx: Option<mpsc::Receiver<ControlMsg>>,
-
-    /// Sender for messages sent by the receiver
-    pdata_tx: mpsc::Sender<TestMsg>,
-    /// Receiver for messages sent by the receiver
-    pdata_rx: Option<mpsc::Receiver<TestMsg>>,
-
     /// Message counter for tracking processed messages
-    counter: CtrMsgCounters,
+    counter: CtrlMsgCounters,
+
+    _pd: PhantomData<PData>,
 }
 
-impl ReceiverTestRuntime {
+/// Data and operations for the test phase of a receiver.
+pub struct TestPhase<PData> {
+    /// Runtime instance
+    rt: tokio::runtime::Runtime,
+    /// Local task set for non-Send futures
+    local_tasks: LocalSet,
+
+    control_sender: Sender<ControlMsg>,
+    receiver: ReceiverWrapper<PData>,
+    counters: CtrlMsgCounters,
+}
+
+/// Data and operations for the validation phase of a receiver.
+pub struct ValidationPhase<PData> {
+    /// Runtime instance
+    rt: tokio::runtime::Runtime,
+    /// Local task set for non-Send futures
+    local_tasks: LocalSet,
+
+    counters: CtrlMsgCounters,
+
+    pdata_receiver: Receiver<PData>,
+
+    /// Join handle for the running the receiver task
+    run_receiver_handle: tokio::task::JoinHandle<()>,
+
+    /// Join handle for the running the test task
+    run_test_handle: tokio::task::JoinHandle<()>,
+}
+
+impl<PData: Clone + Debug + 'static> TestRuntime<PData> {
     /// Creates a new test runtime with channels of the specified capacity.
-    pub fn new(channel_capacity: usize) -> Self {
+    pub fn new() -> Self {
+        let config = ReceiverConfig::new("test_receiver");
         let (rt, local_tasks) = setup_test_runtime();
-        let (control_tx, control_rx) = create_test_channel(channel_capacity);
-        let (pdata_tx, pdata_rx) = create_test_channel(channel_capacity);
 
         Self {
+            config,
             rt,
             local_tasks,
-            control_tx,
-            control_rx: Some(control_rx),
-            pdata_tx,
-            pdata_rx: Some(pdata_rx),
-            counter: CtrMsgCounters::new(),
+            counter: CtrlMsgCounters::new(),
+            _pd: PhantomData,
         }
     }
 
+    /// Returns the current receiver configuration.
+    pub fn config(&self) -> &ReceiverConfig {
+        &self.config
+    }
+
     /// Returns the message counter.
-    pub fn counters(&self) -> CtrMsgCounters {
+    pub fn counters(&self) -> CtrlMsgCounters {
         self.counter.clone()
     }
 
-    /// Starts a receiver with the configured channels.
-    pub fn start_receiver<R>(&mut self, receiver: R)
-    where
-        R: Receiver<PData = TestMsg> + 'static,
-    {
-        let control_rx = self
-            .control_rx
-            .take()
-            .expect("Control channel not initialized");
-        let ctrl_msg_chan = ControlMsgChannel::new(control_rx);
-        let pdata_tx = self.pdata_tx.clone();
+    /// Sets the receiver for the test runtime and returns a test phase.
+    pub fn set_receiver(self, receiver: ReceiverWrapper<PData>) -> TestPhase<PData> {
+        let control_sender = receiver.control_sender();
+        TestPhase {
+            rt: self.rt,
+            local_tasks: self.local_tasks,
+            receiver,
+            control_sender,
+            counters: self.counter,
+        }
+    }
+}
 
-        let receiver = Box::new(receiver);
-        let _ = self.local_tasks.spawn_local(async move {
-            receiver
-                .start(
-                    ctrl_msg_chan,
-                    crate::receiver::EffectHandler::new("test_receiver", pdata_tx),
-                )
+impl<PData: Debug + 'static> TestPhase<PData> {
+    /// Starts the test scenario by executing the provided function with the test context.
+    pub fn run_test<F, Fut>(mut self, f: F) -> ValidationPhase<PData>
+    where
+        F: FnOnce(TestContext) -> Fut + 'static,
+        Fut: Future<Output = ()> + 'static,
+    {
+        let pdata_receiver = self.receiver.take_pdata_receiver();
+        let run_receiver_handle = self.local_tasks.spawn_local(async move {
+            self.receiver
+                .start()
                 .await
                 .expect("Receiver event loop failed");
         });
-    }
 
-    /// Starts the test scenario by executing the provided function with the test context.
-    pub fn start_test<F, Fut>(&mut self, f: F)
-    where
-        F: FnOnce(ReceiverTestContext) -> Fut + 'static,
-        Fut: Future<Output = ()> + 'static,
-    {
-        let context = ReceiverTestContext {
-            control_tx: self.control_tx.clone(),
-            pdata_rx: None,
+        let context = TestContext {
+            control_sender: self.control_sender,
         };
-        let _ = self.local_tasks.spawn_local(async move {
+        let run_test_handle = self.local_tasks.spawn_local(async move {
             f(context).await;
         });
+        ValidationPhase {
+            rt: self.rt,
+            local_tasks: self.local_tasks,
+            counters: self.counters,
+            pdata_receiver,
+            run_receiver_handle,
+            run_test_handle,
+        }
     }
+}
 
+impl<PData> ValidationPhase<PData> {
     /// Runs all spawned tasks to completion and executes the provided future to validate test
     /// expectations.
     ///
@@ -171,18 +245,26 @@ impl ReceiverTestRuntime {
     /// # Returns
     ///
     /// The result of the provided future.
-    pub fn validate<F, Fut, T>(mut self, future_fn: F) -> T
+    pub fn run_validation<F, Fut, T>(self, future_fn: F) -> T
     where
-        F: FnOnce(ReceiverTestContext) -> Fut,
+        F: FnOnce(NotSendValidateContext<PData>) -> Fut,
         Fut: Future<Output = T>,
     {
+        let context = NotSendValidateContext {
+            pdata_receiver: self.pdata_receiver,
+            counters: self.counters,
+        };
+
         // First run all the spawned tasks to completion
         self.rt.block_on(self.local_tasks);
 
-        let context = ReceiverTestContext {
-            control_tx: self.control_tx.clone(),
-            pdata_rx: self.pdata_rx.take(),
-        };
+        self.rt
+            .block_on(self.run_receiver_handle)
+            .expect("Receiver task failed");
+
+        self.rt
+            .block_on(self.run_test_handle)
+            .expect("Test task failed");
 
         // Then run the validation future with the test context
         self.rt.block_on(future_fn(context))
