@@ -12,14 +12,16 @@
 
 use crate::decode::record_message::RecordMessage;
 use crate::error;
-use crate::opentelemetry::{ArrowPayload, ArrowPayloadType, BatchArrowRecords};
-use crate::otlp::metric::metrics_from;
-use crate::otlp::related_data::RelatedData;
+use crate::otlp::logs::logs_from;
+use crate::otlp::logs::related_data::RelatedData as LogsRelatedData;
+use crate::otlp::metrics::{metrics_from, related_data::RelatedData as MetricsRelatedData};
+use crate::proto::opentelemetry::arrow::v1::{ArrowPayload, ArrowPayloadType, BatchArrowRecords};
+use crate::proto::opentelemetry::collector::logs::v1::ExportLogsServiceRequest;
 use crate::proto::opentelemetry::collector::metrics::v1::ExportMetricsServiceRequest;
 use arrow::array::RecordBatch;
 use arrow::error::ArrowError;
 use arrow::ipc::reader::StreamReader;
-use snafu::{ensure, OptionExt, ResultExt};
+use snafu::{OptionExt, ResultExt, ensure};
 use std::collections::HashMap;
 use std::io::Cursor;
 
@@ -48,6 +50,7 @@ impl StreamConsumer {
     }
 }
 
+/// Consumer consumes OTAP `BatchArrowRecords` and converts them into OTLP messages.
 #[derive(Default)]
 pub struct Consumer {
     stream_consumers: HashMap<String, StreamConsumer>,
@@ -103,42 +106,63 @@ impl Consumer {
         Ok(records)
     }
 
-    pub fn consume_batches(
+    /// Consumes all the arrow payloads in the passed OTAP `BatchArrayRecords` and decodes them
+    /// into OTLP messages, then constructs the `ExportMetricsServiceRequest` containing the
+    /// metrics messages
+    pub fn consume_metrics_batches(
         &mut self,
         records: &mut BatchArrowRecords,
     ) -> error::Result<ExportMetricsServiceRequest> {
-        ensure!(!records.arrow_payloads.is_empty(), error::EmptyBatchSnafu);
-
-        let main_record_type = records.arrow_payloads[0].r#type;
-        let payload_type = ArrowPayloadType::try_from(main_record_type).map_err(|_| {
-            error::UnsupportedPayloadTypeSnafu {
-                actual: main_record_type,
-            }
-            .build()
-        })?;
-        match payload_type {
+        match get_main_payload_type(records)? {
             ArrowPayloadType::UnivariateMetrics => {
                 let record_message = self.consume_bar(records)?;
                 let (mut related_data, metric_record) =
-                    RelatedData::from_record_messages(&record_message)?;
+                    MetricsRelatedData::from_record_messages(&record_message)?;
                 let metric_rec_idx = metric_record.context(error::MetricRecordNotFoundSnafu)?;
                 metrics_from(&record_message[metric_rec_idx].record, &mut related_data)
             }
-
-            ArrowPayloadType::Logs => error::UnsupportedPayloadTypeSnafu {
-                actual: main_record_type,
-            }
-            .fail(),
-            ArrowPayloadType::Spans => error::UnsupportedPayloadTypeSnafu {
-                actual: main_record_type,
-            }
-            .fail(),
-            _ => error::UnsupportedPayloadTypeSnafu {
+            main_record_type => error::UnsupportedPayloadTypeSnafu {
                 actual: main_record_type,
             }
             .fail(),
         }
     }
+
+    /// Consumes all the arrow payloads in the passed OTAP `BatchArrayRecords` and decodes them
+    /// into OTLP messages, then constructs the `ExportLogsServiceRequest` containing the
+    /// logs messages
+    pub fn consume_logs_batches(
+        &mut self,
+        records: &mut BatchArrowRecords,
+    ) -> error::Result<ExportLogsServiceRequest> {
+        match get_main_payload_type(records)? {
+            ArrowPayloadType::Logs => {
+                let record_message = self.consume_bar(records)?;
+                let (mut related_data, log_record) =
+                    LogsRelatedData::from_record_messages(&record_message)?;
+                let log_rec_idx = log_record.context(error::LogRecordNotFoundSnafu)?;
+                logs_from(&record_message[log_rec_idx].record, &mut related_data)
+            }
+            main_record_type => error::UnsupportedPayloadTypeSnafu {
+                actual: main_record_type,
+            }
+            .fail(),
+        }
+    }
+}
+
+/// Get the main logs, metrics, or traces from a received BatchArrowRecords message.
+fn get_main_payload_type(records: &BatchArrowRecords) -> error::Result<ArrowPayloadType> {
+    ensure!(!records.arrow_payloads.is_empty(), error::EmptyBatchSnafu);
+
+    // Per the specification, the main record type is the first payload
+    let main_record_type = records.arrow_payloads[0].r#type;
+    ArrowPayloadType::try_from(main_record_type).map_err(|_| {
+        error::UnsupportedPayloadTypeSnafu {
+            actual: main_record_type,
+        }
+        .build()
+    })
 }
 
 #[cfg(test)]
