@@ -11,12 +11,12 @@
 // limitations under the License.
 
 use crate::otlp::attributes::store::AttributeValueType;
-use crate::otlp::metric::MetricType;
+use crate::otlp::metrics::MetricType;
 use arrow::datatypes::DataType;
 use arrow::error::ArrowError;
 use num_enum::TryFromPrimitiveError;
 use snafu::{Location, Snafu};
-use std::backtrace::Backtrace;
+use std::{backtrace::Backtrace, num::TryFromIntError};
 
 pub type Result<T> = std::result::Result<T, Error>;
 
@@ -61,9 +61,33 @@ pub enum Error {
         location: Location,
     },
 
-    #[snafu(display("Currently attribute store value type: {} is not supported", type_name))]
-    UnsupportedAttributeValue {
-        type_name: String,
+    #[snafu(display("Invalid bytes for serialized attribute value"))]
+    InvalidSerializedAttributeBytes {
+        source: ciborium::de::Error<std::io::Error>,
+        #[snafu(implicit)]
+        location: Location,
+    },
+
+    #[snafu(display("Invalid serialized integer attribute value"))]
+    InvalidSerializedIntAttributeValue {
+        source: TryFromIntError,
+        #[snafu(implicit)]
+        location: Location,
+    },
+
+    #[snafu(display(
+        "Invalid serialized map key type, expected: String, actual: {:?}",
+        actual
+    ))]
+    InvalidSerializedMapKeyType {
+        actual: ciborium::Value,
+        #[snafu(implicit)]
+        location: Location,
+    },
+
+    #[snafu(display("Serialized attribute {:?} is not supported", actual))]
+    UnsupportedSerializedAttributeValue {
+        actual: ciborium::Value,
         #[snafu(implicit)]
         location: Location,
     },
@@ -96,9 +120,13 @@ pub enum Error {
         location: Location,
     },
 
-    #[snafu(display("Invalid List array data type, expect {}, actual {}", expect, actual))]
+    #[snafu(display(
+        "Invalid List array data type, expect one of {:?}, actual {}",
+        expect_oneof,
+        actual
+    ))]
     InvalidListArray {
-        expect: DataType,
+        expect_oneof: Vec<DataType>,
         actual: DataType,
         #[snafu(implicit)]
         location: Location,
@@ -133,8 +161,38 @@ pub enum Error {
         location: Location,
     },
 
+    #[snafu(display("Log record not found"))]
+    LogRecordNotFound {
+        #[snafu(implicit)]
+        location: Location,
+    },
+
     #[snafu(display("Metric record not found"))]
     MetricRecordNotFound {
+        #[snafu(implicit)]
+        location: Location,
+    },
+
+    #[snafu(display(
+        "Unsupported dictionary key type, expect one of {:?}, actual {}",
+        expect_oneof,
+        actual
+    ))]
+    UnsupportedDictionaryKeyType {
+        expect_oneof: Vec<DataType>,
+        actual: DataType,
+        #[snafu(implicit)]
+        location: Location,
+    },
+
+    #[snafu(display(
+        "Unsupported dictionary value type. expect {:?}, actual {}",
+        expect_oneof,
+        actual
+    ))]
+    UnsupportedDictionaryValueType {
+        expect_oneof: Vec<DataType>,
+        actual: DataType,
         #[snafu(implicit)]
         location: Location,
     },
