@@ -10,7 +10,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#![feature(let_chains)]
+//! This crate contains the rust implementation of the OTEL Arrow Protocol.
+//! It contains code for decoding OTAP Arrow record batches into OTLP protos,
+//! and encoding OTLP protos into OTAP Messages. It also contains
+//! the rust implementation of pdata.
 
 #[allow(dead_code)]
 pub(crate) mod arrays;
@@ -21,26 +24,10 @@ mod otlp;
 mod schema;
 #[cfg(test)]
 mod test_util;
+#[cfg(test)]
+mod validation;
 
 pub mod pdata;
 pub mod proto;
-
-#[path = ""]
-pub mod opentelemetry {
-    pub use proto::*;
-    pub mod proto {
-        pub use crate::opentelemetry::proto::arrow::arrow_metrics_service_client as metrics_client;
-        pub use crate::opentelemetry::proto::arrow::arrow_metrics_service_server as metrics_server;
-        pub use crate::opentelemetry::proto::arrow::{
-            ArrowPayload, ArrowPayloadType, BatchArrowRecords, BatchStatus,
-        };
-        pub use metrics_client::ArrowMetricsServiceClient;
-        pub use metrics_server::{ArrowMetricsService, ArrowMetricsServiceServer};
-
-        #[allow(clippy::all)]
-        #[path = "opentelemetry.proto.experimental.arrow.v1.rs"]
-        pub(crate) mod arrow;
-    }
-}
 
 pub use decode::decoder::Consumer;
