@@ -10,9 +10,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod exemplar;
-pub mod metric;
-
 pub mod attributes;
-pub mod data_points;
-pub mod related_data;
+pub mod logs;
+pub mod metrics;
+
+mod common;
