@@ -12,14 +12,21 @@
 
 // https://github.com/open-telemetry/otel-arrow/blob/985aa1500a012859cec44855e187eacf46eda7c8/pkg/otel/common/arrow/attributes.go#L40
 
-use crate::otlp::attributes::parent_id::{ParentId, ParentIdEncoding};
+use crate::otlp::attributes::parent_id::ParentId;
 use crate::proto::opentelemetry::common::v1::any_value;
 
 pub type Attrs16ParentIdDecoder = AttrsParentIdDecoder<u16>;
 pub type Attrs32ParentIdDecoder = AttrsParentIdDecoder<u32>;
 
+// AttrsParentIdDecoder implements parent_id decoding for attribute
+// sets.  The parent_id in this case is the entity which refers to the
+// set of attributes (e.g., a resource, a scope, a metric) contained
+// in a RecordBatch.
+//
+// Phase 1 note: there were several experimental encoding schemes
+// tested.  Two schemes named "ParentIdDeltaEncoding",
+// "ParentIdNoEncoding" have been removed.
 pub struct AttrsParentIdDecoder<T> {
-    encoding_type: ParentIdEncoding,
     prev_parent_id: T,
     prev_key: Option<String>,
     prev_value: Option<any_value::Value>,
@@ -31,7 +38,6 @@ where
 {
     fn default() -> Self {
         Self {
-            encoding_type: ParentIdEncoding::ParentIdDeltaGroupEncoding,
             prev_parent_id: T::default(),
             prev_key: None,
             prev_value: None,
@@ -44,29 +50,15 @@ where
     T: ParentId,
 {
     pub fn decode(&mut self, delta_or_parent_id: T, key: &str, value: &any_value::Value) -> T {
-        match self.encoding_type {
-            // Plain encoding
-            ParentIdEncoding::ParentIdNoEncoding => delta_or_parent_id,
-            // Simply delta
-            ParentIdEncoding::ParentIdDeltaEncoding => {
-                let decode_parent_id = self.prev_parent_id.add(delta_or_parent_id);
-                self.prev_parent_id = decode_parent_id;
-                decode_parent_id
-            }
-            // Key-value scoped delta.
-            ParentIdEncoding::ParentIdDeltaGroupEncoding => {
-                if self.prev_key.as_deref() == Some(key) && self.prev_value.as_ref() == Some(value)
-                {
-                    let parent_id = self.prev_parent_id.add(delta_or_parent_id);
-                    self.prev_parent_id = parent_id;
-                    parent_id
-                } else {
-                    self.prev_key = Some(key.to_string());
-                    self.prev_value = Some(value.clone());
-                    self.prev_parent_id = delta_or_parent_id;
-                    delta_or_parent_id
-                }
-            }
+        if self.prev_key.as_deref() == Some(key) && self.prev_value.as_ref() == Some(value) {
+            let parent_id = self.prev_parent_id.add(delta_or_parent_id);
+            self.prev_parent_id = parent_id;
+            parent_id
+        } else {
+            self.prev_key = Some(key.to_string());
+            self.prev_value = Some(value.clone());
+            self.prev_parent_id = delta_or_parent_id;
+            delta_or_parent_id
         }
     }
 }
