@@ -10,30 +10,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::arrays::NullableArrayAccessor;
 use crate::otlp::attributes::decoder::{
     Attrs16ParentIdDecoder, Attrs32ParentIdDecoder, AttrsParentIdDecoder,
 };
-use arrow::array::{UInt16Array, UInt32Array};
-use arrow::datatypes::DataType;
-use num_enum::TryFromPrimitive;
+use arrow::datatypes::{UInt16Type, UInt32Type};
 use std::hash::Hash;
 use std::ops::{Add, AddAssign};
 
 pub trait ParentId: Copy + Hash + Eq + Default + Add<Output = Self> + AddAssign {
-    type Array: NullableArrayAccessor<Native = Self> + 'static;
-
-    fn arrow_data_type() -> DataType;
+    type ArrayType;
 
     fn new_decoder() -> AttrsParentIdDecoder<Self>;
 }
 
 impl ParentId for u16 {
-    type Array = UInt16Array;
-
-    fn arrow_data_type() -> DataType {
-        DataType::UInt16
-    }
+    type ArrayType = UInt16Type;
 
     fn new_decoder() -> AttrsParentIdDecoder<Self> {
         Attrs16ParentIdDecoder::default()
@@ -41,28 +32,9 @@ impl ParentId for u16 {
 }
 
 impl ParentId for u32 {
-    type Array = UInt32Array;
-
-    fn arrow_data_type() -> DataType {
-        DataType::UInt32
-    }
+    type ArrayType = UInt32Type;
 
     fn new_decoder() -> AttrsParentIdDecoder<Self> {
         Attrs32ParentIdDecoder::default()
     }
-}
-
-#[allow(clippy::enum_variant_names)]
-#[derive(Copy, Clone, Eq, PartialEq, Debug, TryFromPrimitive)]
-#[repr(u8)]
-pub enum ParentIdEncoding {
-    /// ParentIdNoEncoding stores the parent ID as is.
-    ParentIdNoEncoding = 0,
-    /// ParentIdDeltaEncoding stores the parent ID as a delta from the previous
-    /// parent ID.
-    ParentIdDeltaEncoding = 1,
-    /// ParentIdDeltaGroupEncoding stores the parent ID as a delta from the
-    /// previous parent ID in the same group. A group is defined by the
-    /// combination Key and Value.
-    ParentIdDeltaGroupEncoding = 2,
 }
