@@ -14,26 +14,61 @@
 
 pub mod otlp;
 
-// TODO use the OTel-Rust API definitions, if such a dependency.
+// Note that these types are placeholders, we probably want to share
+// these definitions as well as the Prost/Tonic generation with the
+// OTel-Rust SDK where they are already defined. To avoid coordinating
+// these repositories in the short term, we provide definitions for
+// TraceID and SpanID.
+//
+// In particular, the OTel specification has careful words about how
+// to format and parse these two fields, which are non-standard with
+// respect to JSON, and the OTel-Rust SDK implements this aspect of
+// the spec.
 
-// TODO: otherwise, specialize TraceID and SpanID for [u8; 8 or 16].
-
-// TODO: also, define From<> instead of Into<> in ../proto/mod.rs
-
+/// TraceID identifier of a Trace
 #[derive(Clone, Copy, Debug)]
 pub struct TraceID([u8; 16]);
 
-impl Into<Vec<u8>> for TraceID {
-    fn into(self) -> Vec<u8> {
-        self.0.to_vec()
+impl TraceID {
+    /// creates a new instance of the TraceID by copying the bytes
+    #[must_use]
+    pub fn new(value: &[u8; 16]) -> TraceID {
+        TraceID(*value)
     }
 }
 
+impl From<[u8; 16]> for TraceID {
+    fn from(tid: [u8; 16]) -> Self {
+        TraceID(tid)
+    }
+}
+
+impl From<TraceID> for Vec<u8> {
+    fn from(tid: TraceID) -> Self {
+        tid.0.to_vec()
+    }
+}
+
+/// SpanID identifier of a Span
 #[derive(Clone, Copy, Debug)]
 pub struct SpanID([u8; 8]);
 
-impl Into<Vec<u8>> for SpanID {
-    fn into(self) -> Vec<u8> {
-        self.0.to_vec()
+impl SpanID {
+    /// creates a new instance of the SpanID by copying the bytes
+    #[must_use]
+    pub fn new(value: &[u8; 8]) -> SpanID {
+        SpanID(*value)
+    }
+}
+
+impl From<SpanID> for Vec<u8> {
+    fn from(sid: SpanID) -> Self {
+        sid.0.to_vec()
+    }
+}
+
+impl From<[u8; 8]> for SpanID {
+    fn from(sid: [u8; 8]) -> Self {
+        SpanID(sid)
     }
 }
