@@ -11,8 +11,8 @@
 // limitations under the License.
 
 use proc_macro::TokenStream;
-use quote::{quote, ToTokens};
-use syn::{parse_macro_input, DeriveInput};
+use quote::{ToTokens, quote};
+use syn::{DeriveInput, parse_macro_input};
 
 /// Attribute macro for associating the OTLP protocol buffer fully
 /// qualified type name.
@@ -98,7 +98,7 @@ pub fn derive_otlp_message(input: TokenStream) -> TokenStream {
 
     // If there are no fields, it's either an empty message or an enum,
     // either way should not be listed, no builder is needed.
-    if struct_fields.len() == 0 {
+    if struct_fields.is_empty() {
         panic!("message with empty fields")
     }
 
@@ -155,9 +155,7 @@ pub fn derive_otlp_message(input: TokenStream) -> TokenStream {
                 let field_path = format!("{}.{}", type_name, ident_str);
                 let is_param = param_names.contains(&ident_str.as_str());
                 let is_optional = is_optional(field);
-                let is_oneof = oneof_mapping
-                    .map(|x| x.0.to_string() == field_path)
-                    .unwrap_or(false);
+                let is_oneof = oneof_mapping.map(|x| *x.0 == field_path).unwrap_or(false);
 
                 // Process type information
                 let (inner_type, is_optional_extraction_ok) = if is_optional {
@@ -204,7 +202,10 @@ pub fn derive_otlp_message(input: TokenStream) -> TokenStream {
         .map(|param_name| {
             fields_original
                 .iter()
-                .find(|info| info.is_param && info.ident.to_string() == *param_name)
+                .find(|info| {
+                    let ident = info.ident.to_string();
+                    info.is_param && ident == *param_name
+                })
                 .unwrap()
         })
         .collect();
@@ -381,8 +382,8 @@ pub fn derive_otlp_message(input: TokenStream) -> TokenStream {
 
             // Generate a constructor for each oneof case
             oneof_mapping.1.iter().map(|case| {
-                let case_type = syn::parse_str::<syn::Type>(&case.type_param).unwrap();
-                let variant_path = syn::parse_str::<syn::Expr>(&case.value_variant).unwrap();
+                let case_type = syn::parse_str::<syn::Type>(case.type_param).unwrap();
+                let variant_path = syn::parse_str::<syn::Expr>(case.value_variant).unwrap();
                 let suffix = format!("_{}", case.name);
 
                 // Duplicate the param bounds, assignments; param decls unchanged.
