@@ -162,6 +162,43 @@ pub static REQUIRED_PARAMS: LazyLock<HashMap<&'static str, Vec<&'static str>>> =
                 "time_unix_nano",
                 "value",
             ]),
+            // Service
+            (
+                "opentelemetry.proto.collector.logs.v1.ExportLogsServiceRequest",
+                vec!["resource_logs"],
+            ),
+            (
+                "opentelemetry.proto.collector.logs.v1.ExportLogsServiceResponse",
+                vec!["partial_success"],
+            ),
+            (
+                "opentelemetry.proto.collector.logs.v1.ExportLogsPartialSuccess",
+                vec!["rejected_log_records"],
+            ),
+            (
+                "opentelemetry.proto.collector.metrics.v1.ExportMetricsServiceRequest",
+                vec!["resource_metrics"],
+            ),
+            (
+                "opentelemetry.proto.collector.metrics.v1.ExportMetricsServiceResponse",
+                vec!["partial_success"],
+            ),
+            (
+                "opentelemetry.proto.collector.metrics.v1.ExportMetricsPartialSuccess",
+                vec!["rejected_data_points"],
+            ),
+            (
+                "opentelemetry.proto.collector.trace.v1.ExportTraceServiceRequest",
+                vec!["resource_spans"],
+            ),
+            (
+                "opentelemetry.proto.collector.trace.v1.ExportTraceServiceResponse",
+                vec!["partial_success"],
+            ),
+            (
+                "opentelemetry.proto.collector.trace.v1.ExportTracePartialSuccess",
+                vec!["rejected_spans"],
+            ),
         ])
     });
 
@@ -256,10 +293,8 @@ pub static FIELD_TYPE_OVERRIDES: LazyLock<HashMap<&'static str, EnumField>> = La
 /// This is the entry point from build.rs where we configure prost/tonic.
 pub fn add_type_attributes(mut builder: tonic_build::Builder) -> tonic_build::Builder {
     for (name, _) in REQUIRED_PARAMS.iter() {
-        builder = builder.type_attribute(
-            name,
-            &format!(r#"#[crate::pdata::otlp::qualified("{}")]"#, name),
-        );
+        let attr = format!(r#"#[crate::pdata::otlp::qualified("{}")]"#, name);
+        builder = builder.type_attribute(name, attr);
         builder = builder.type_attribute(name, r#"#[derive(crate::pdata::otlp::Message)]"#);
     }
     builder
