@@ -1,49 +1,21 @@
 // This file is @generated by prost-build.
-#[derive(Clone, PartialEq, ::prost::Message)]
+#[crate::pdata::otlp::qualified("opentelemetry.proto.collector.logs.v1.ExportLogsServiceRequest")]
+#[derive(crate::pdata::otlp::Message, Clone, PartialEq, ::prost::Message)]
 pub struct ExportLogsServiceRequest {
-    /// An array of ResourceLogs.
-    /// For data coming from a single resource this array will typically contain one
-    /// element. Intermediary nodes (such as OpenTelemetry Collector) that receive
-    /// data from multiple origins typically batch the data before forwarding further and
-    /// in that case this array will contain multiple elements.
     #[prost(message, repeated, tag = "1")]
     pub resource_logs: ::prost::alloc::vec::Vec<super::super::super::logs::v1::ResourceLogs>,
 }
-#[derive(Clone, PartialEq, ::prost::Message)]
+#[crate::pdata::otlp::qualified("opentelemetry.proto.collector.logs.v1.ExportLogsServiceResponse")]
+#[derive(crate::pdata::otlp::Message, Clone, PartialEq, ::prost::Message)]
 pub struct ExportLogsServiceResponse {
-    /// The details of a partially successful export request.
-    ///
-    /// If the request is only partially accepted
-    /// (i.e. when the server accepts only parts of the data and rejects the rest)
-    /// the server MUST initialize the `partial_success` field and MUST
-    /// set the `rejected_<signal>` with the number of items it rejected.
-    ///
-    /// Servers MAY also make use of the `partial_success` field to convey
-    /// warnings/suggestions to senders even when the request was fully accepted.
-    /// In such cases, the `rejected_<signal>` MUST have a value of `0` and
-    /// the `error_message` MUST be non-empty.
-    ///
-    /// A `partial_success` message with an empty value (rejected_<signal> = 0 and
-    /// `error_message` = "") is equivalent to it not being set/present. Senders
-    /// SHOULD interpret it the same way as in the full success case.
     #[prost(message, optional, tag = "1")]
     pub partial_success: ::core::option::Option<ExportLogsPartialSuccess>,
 }
-#[derive(Clone, PartialEq, ::prost::Message)]
+#[crate::pdata::otlp::qualified("opentelemetry.proto.collector.logs.v1.ExportLogsPartialSuccess")]
+#[derive(crate::pdata::otlp::Message, Clone, PartialEq, ::prost::Message)]
 pub struct ExportLogsPartialSuccess {
-    /// The number of rejected log records.
-    ///
-    /// A `rejected_<signal>` field holding a `0` value indicates that the
-    /// request was fully accepted.
     #[prost(int64, tag = "1")]
     pub rejected_log_records: i64,
-    /// A developer-facing human-readable message in English. It should be used
-    /// either to explain why the server rejected parts of the data during a partial
-    /// success or to convey warnings/suggestions during a full success. The message
-    /// should offer guidance on how users can address such issues.
-    ///
-    /// error_message is an optional field. An error_message with an empty value
-    /// is equivalent to it not being set.
     #[prost(string, tag = "2")]
     pub error_message: ::prost::alloc::string::String,
 }
@@ -79,7 +51,7 @@ pub mod logs_service_client {
     }
     impl<T> LogsServiceClient<T>
     where
-        T: tonic::client::GrpcService<tonic::body::BoxBody>,
+        T: tonic::client::GrpcService<tonic::body::Body>,
         T::Error: Into<StdError>,
         T::ResponseBody: Body<Data = Bytes> + std::marker::Send + 'static,
         <T::ResponseBody as Body>::Error: Into<StdError> + std::marker::Send,
@@ -100,12 +72,12 @@ pub mod logs_service_client {
             F: tonic::service::Interceptor,
             T::ResponseBody: Default,
             T: tonic::codegen::Service<
-                http::Request<tonic::body::BoxBody>,
-                Response = http::Response<
-                    <T as tonic::client::GrpcService<tonic::body::BoxBody>>::ResponseBody,
+                    http::Request<tonic::body::Body>,
+                    Response = http::Response<
+                        <T as tonic::client::GrpcService<tonic::body::Body>>::ResponseBody,
+                    >,
                 >,
-            >,
-            <T as tonic::codegen::Service<http::Request<tonic::body::BoxBody>>>::Error:
+            <T as tonic::codegen::Service<http::Request<tonic::body::Body>>>::Error:
                 Into<StdError> + std::marker::Send + std::marker::Sync,
         {
             LogsServiceClient::new(InterceptedService::new(inner, interceptor))
@@ -246,7 +218,7 @@ pub mod logs_service_server {
         B: Body + std::marker::Send + 'static,
         B::Error: Into<StdError> + std::marker::Send + 'static,
     {
-        type Response = http::Response<tonic::body::BoxBody>;
+        type Response = http::Response<tonic::body::Body>;
         type Error = std::convert::Infallible;
         type Future = BoxFuture<Self::Response, Self::Error>;
         fn poll_ready(
@@ -299,7 +271,7 @@ pub mod logs_service_server {
                     Box::pin(fut)
                 }
                 _ => Box::pin(async move {
-                    let mut response = http::Response::new(empty_body());
+                    let mut response = http::Response::new(tonic::body::Body::default());
                     let headers = response.headers_mut();
                     headers.insert(
                         tonic::Status::GRPC_STATUS,
