@@ -1,40 +1,22 @@
 // This file is @generated by prost-build.
-/// A message sent by an exporter to a collector containing a batch of Arrow
-/// records.
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct BatchArrowRecords {
-    /// \[mandatory\] Batch ID. Must be unique in the context of the stream.
     #[prost(int64, tag = "1")]
     pub batch_id: i64,
-    /// \[mandatory\] A collection of payloads containing the data of the batch.
     #[prost(message, repeated, tag = "2")]
     pub arrow_payloads: ::prost::alloc::vec::Vec<ArrowPayload>,
-    /// \[optional\] Headers associated with this batch, encoded using hpack.
     #[prost(bytes = "vec", tag = "3")]
     pub headers: ::prost::alloc::vec::Vec<u8>,
 }
-/// Represents a batch of OTel Arrow entities.
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct ArrowPayload {
-    /// \[mandatory\] A canonical ID representing the schema of the Arrow Record.
-    /// This ID is used on the consumer side to determine the IPC reader to use
-    /// for interpreting the corresponding record. For any NEW `schema_id`, the
-    /// consumer must:
-    /// 1) close the current IPC reader,
-    /// 2) create a new IPC reader in order to interpret the new schema,
-    /// dictionaries, and corresponding data.
     #[prost(string, tag = "1")]
     pub schema_id: ::prost::alloc::string::String,
-    /// \[mandatory\] Type of the OTel Arrow payload.
     #[prost(enumeration = "ArrowPayloadType", tag = "2")]
     pub r#type: i32,
-    /// \[mandatory\] Serialized Arrow Record Batch
-    /// For a description of the Arrow IPC format see:
-    /// <https://arrow.apache.org/docs/format/Columnar.html#serialization-and-interprocess-communication-ipc>
     #[prost(bytes = "vec", tag = "3")]
     pub record: ::prost::alloc::vec::Vec<u8>,
 }
-/// A message sent by a Collector to the exporter that opened the data stream.
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct BatchStatus {
     #[prost(int64, tag = "1")]
@@ -44,19 +26,12 @@ pub struct BatchStatus {
     #[prost(string, tag = "3")]
     pub status_message: ::prost::alloc::string::String,
 }
-/// Enumeration of all the OTel Arrow payload types currently supported by the
-/// OTel Arrow protocol.
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration)]
 #[repr(i32)]
 pub enum ArrowPayloadType {
     Unknown = 0,
-    /// A payload representing a collection of resource attributes.
     ResourceAttrs = 1,
-    /// A payload representing a collection of scope attributes.
     ScopeAttrs = 2,
-    /// A set of payloads representing a collection of metrics.
-    ///
-    /// Univariate metric payload
     UnivariateMetrics = 10,
     NumberDataPoints = 11,
     SummaryDataPoints = 12,
@@ -72,12 +47,9 @@ pub enum ArrowPayloadType {
     NumberDpExemplarAttrs = 22,
     HistogramDpExemplarAttrs = 23,
     ExpHistogramDpExemplarAttrs = 24,
-    /// Multivariate metric payload
     MultivariateMetrics = 25,
-    /// A set of payloads representing a collection of logs.
     Logs = 30,
     LogAttrs = 31,
-    /// A set of payloads representing a collection of traces.
     Spans = 40,
     SpanAttrs = 41,
     SpanEvents = 42,
@@ -155,8 +127,6 @@ impl ArrowPayloadType {
         }
     }
 }
-/// StatusCode carries certain known meanings in Arrow.  Values match
-/// the gRPC code space, see <https://grpc.io/docs/guides/status-codes/.>
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration)]
 #[repr(i32)]
 pub enum StatusCode {
@@ -237,7 +207,7 @@ pub mod arrow_traces_service_client {
     }
     impl<T> ArrowTracesServiceClient<T>
     where
-        T: tonic::client::GrpcService<tonic::body::BoxBody>,
+        T: tonic::client::GrpcService<tonic::body::Body>,
         T::Error: Into<StdError>,
         T::ResponseBody: Body<Data = Bytes> + std::marker::Send + 'static,
         <T::ResponseBody as Body>::Error: Into<StdError> + std::marker::Send,
@@ -258,12 +228,12 @@ pub mod arrow_traces_service_client {
             F: tonic::service::Interceptor,
             T::ResponseBody: Default,
             T: tonic::codegen::Service<
-                http::Request<tonic::body::BoxBody>,
-                Response = http::Response<
-                    <T as tonic::client::GrpcService<tonic::body::BoxBody>>::ResponseBody,
+                    http::Request<tonic::body::Body>,
+                    Response = http::Response<
+                        <T as tonic::client::GrpcService<tonic::body::Body>>::ResponseBody,
+                    >,
                 >,
-            >,
-            <T as tonic::codegen::Service<http::Request<tonic::body::BoxBody>>>::Error:
+            <T as tonic::codegen::Service<http::Request<tonic::body::Body>>>::Error:
                 Into<StdError> + std::marker::Send + std::marker::Sync,
         {
             ArrowTracesServiceClient::new(InterceptedService::new(inner, interceptor))
@@ -352,7 +322,7 @@ pub mod arrow_logs_service_client {
     }
     impl<T> ArrowLogsServiceClient<T>
     where
-        T: tonic::client::GrpcService<tonic::body::BoxBody>,
+        T: tonic::client::GrpcService<tonic::body::Body>,
         T::Error: Into<StdError>,
         T::ResponseBody: Body<Data = Bytes> + std::marker::Send + 'static,
         <T::ResponseBody as Body>::Error: Into<StdError> + std::marker::Send,
@@ -373,12 +343,12 @@ pub mod arrow_logs_service_client {
             F: tonic::service::Interceptor,
             T::ResponseBody: Default,
             T: tonic::codegen::Service<
-                http::Request<tonic::body::BoxBody>,
-                Response = http::Response<
-                    <T as tonic::client::GrpcService<tonic::body::BoxBody>>::ResponseBody,
+                    http::Request<tonic::body::Body>,
+                    Response = http::Response<
+                        <T as tonic::client::GrpcService<tonic::body::Body>>::ResponseBody,
+                    >,
                 >,
-            >,
-            <T as tonic::codegen::Service<http::Request<tonic::body::BoxBody>>>::Error:
+            <T as tonic::codegen::Service<http::Request<tonic::body::Body>>>::Error:
                 Into<StdError> + std::marker::Send + std::marker::Sync,
         {
             ArrowLogsServiceClient::new(InterceptedService::new(inner, interceptor))
@@ -467,7 +437,7 @@ pub mod arrow_metrics_service_client {
     }
     impl<T> ArrowMetricsServiceClient<T>
     where
-        T: tonic::client::GrpcService<tonic::body::BoxBody>,
+        T: tonic::client::GrpcService<tonic::body::Body>,
         T::Error: Into<StdError>,
         T::ResponseBody: Body<Data = Bytes> + std::marker::Send + 'static,
         <T::ResponseBody as Body>::Error: Into<StdError> + std::marker::Send,
@@ -488,12 +458,12 @@ pub mod arrow_metrics_service_client {
             F: tonic::service::Interceptor,
             T::ResponseBody: Default,
             T: tonic::codegen::Service<
-                http::Request<tonic::body::BoxBody>,
-                Response = http::Response<
-                    <T as tonic::client::GrpcService<tonic::body::BoxBody>>::ResponseBody,
+                    http::Request<tonic::body::Body>,
+                    Response = http::Response<
+                        <T as tonic::client::GrpcService<tonic::body::Body>>::ResponseBody,
+                    >,
                 >,
-            >,
-            <T as tonic::codegen::Service<http::Request<tonic::body::BoxBody>>>::Error:
+            <T as tonic::codegen::Service<http::Request<tonic::body::Body>>>::Error:
                 Into<StdError> + std::marker::Send + std::marker::Sync,
         {
             ArrowMetricsServiceClient::new(InterceptedService::new(inner, interceptor))
@@ -639,7 +609,7 @@ pub mod arrow_traces_service_server {
         B: Body + std::marker::Send + 'static,
         B::Error: Into<StdError> + std::marker::Send + 'static,
     {
-        type Response = http::Response<tonic::body::BoxBody>;
+        type Response = http::Response<tonic::body::Body>;
         type Error = std::convert::Infallible;
         type Future = BoxFuture<Self::Response, Self::Error>;
         fn poll_ready(
@@ -695,7 +665,7 @@ pub mod arrow_traces_service_server {
                     Box::pin(fut)
                 }
                 _ => Box::pin(async move {
-                    let mut response = http::Response::new(empty_body());
+                    let mut response = http::Response::new(tonic::body::Body::default());
                     let headers = response.headers_mut();
                     headers.insert(
                         tonic::Status::GRPC_STATUS,
@@ -815,7 +785,7 @@ pub mod arrow_logs_service_server {
         B: Body + std::marker::Send + 'static,
         B::Error: Into<StdError> + std::marker::Send + 'static,
     {
-        type Response = http::Response<tonic::body::BoxBody>;
+        type Response = http::Response<tonic::body::Body>;
         type Error = std::convert::Infallible;
         type Future = BoxFuture<Self::Response, Self::Error>;
         fn poll_ready(
@@ -871,7 +841,7 @@ pub mod arrow_logs_service_server {
                     Box::pin(fut)
                 }
                 _ => Box::pin(async move {
-                    let mut response = http::Response::new(empty_body());
+                    let mut response = http::Response::new(tonic::body::Body::default());
                     let headers = response.headers_mut();
                     headers.insert(
                         tonic::Status::GRPC_STATUS,
@@ -991,7 +961,7 @@ pub mod arrow_metrics_service_server {
         B: Body + std::marker::Send + 'static,
         B::Error: Into<StdError> + std::marker::Send + 'static,
     {
-        type Response = http::Response<tonic::body::BoxBody>;
+        type Response = http::Response<tonic::body::Body>;
         type Error = std::convert::Infallible;
         type Future = BoxFuture<Self::Response, Self::Error>;
         fn poll_ready(
@@ -1047,7 +1017,7 @@ pub mod arrow_metrics_service_server {
                     Box::pin(fut)
                 }
                 _ => Box::pin(async move {
-                    let mut response = http::Response::new(empty_body());
+                    let mut response = http::Response::new(tonic::body::Body::default());
                     let headers = response.headers_mut();
                     headers.insert(
                         tonic::Status::GRPC_STATUS,
